+++ common/src/main/java/com/selfdot/cobblemontrainers/util/DataKeys.java
@@ -27,6 +27,7 @@ public class DataKeys {
     public static final String UNGROUPED = "Ungrouped";
 
     public static final String CONFIG_XP_ENABLED = "xpEnabled";
+    public static final String CONFIG_COMMAND_EXECUTOR = "commandExecutor";
 
     public static final String PLAYER_PARTY_MAXIMUM_LEVEL = "partyMaximumLevel";
     public static final String PLAYER_DEFEAT_REQUIRED_TRAINERS = "defeatRequiredTrainers";